requirements.txt
Flask==3.0.3
openai>=1.0.0
mysql-connector-python==8.0.33
pyodbc>=5.0.1
python-dotenv>=1.0.0
streamlit==1.29.0
pandas>=1.5.0
pytest>=7.0.0
